requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
matplotlib
pandas
pydicom
numpy
python-multipart
